+++ src/headless/core.js
@@ -47,6 +47,7 @@ import {
 dayjs.extend(advancedFormat);
 
 // Add Strophe Namespaces
+Strophe.addNamespace('ACTIVITY', 'http://jabber.org/protocol/activity');
 Strophe.addNamespace('CARBONS', 'urn:xmpp:carbons:2');
 Strophe.addNamespace('CHATSTATES', 'http://jabber.org/protocol/chatstates');
 Strophe.addNamespace('CSI', 'urn:xmpp:csi:0');